deleted module and related files along with occurrences
