+++ README.md
@@ -79,8 +79,6 @@ High-level components for building browser(-like) apps.
 
 * 🔵 [**State**](components/browser/state/README.md) - Component for maintaining the centralized state of the browser and its components.
 
-* 🔵 [**Storage-Memory**](components/browser/storage-memory/README.md) - An in-memory implementation of browser storage.
-
 * 🔵 [**Storage-Sync**](components/browser/storage-sync/README.md) - A syncable implementation of browser storage backed by [application-services' Places lib](https://github.com/mozilla/application-services).
 
 * ⚪ [**Tabstray**](components/browser/tabstray/README.md) - A customizable tabs tray for browsers.
