--- components/browser/storage-memory/src/main/AndroidManifest.xml
@@ -1,5 +1,0 @@
-<!-- This Source Code Form is subject to the terms of the Mozilla Public
-   - License, v. 2.0. If a copy of the MPL was not distributed with this
-   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="mozilla.components.browser.storage.memory" />
