--- components/browser/storage-memory/src/test/resources/robolectric.properties
@@ -1,1 +1,0 @@
-sdk=28
