+++ components/feature/awesomebar/build.gradle
@@ -42,7 +42,6 @@ dependencies {
     implementation Dependencies.kotlin_stdlib
 
     testImplementation project(':support-test')
-    testImplementation project(':browser-storage-memory')
     testImplementation project(':lib-fetch-httpurlconnection')
 
     testImplementation Dependencies.androidx_test_core
