+++ components/feature/toolbar/build.gradle
@@ -44,7 +44,6 @@ dependencies {
     implementation Dependencies.kotlin_coroutines
 
     testImplementation project(':support-test')
-    testImplementation project(':browser-storage-memory')
 
     testImplementation Dependencies.androidx_test_core
     testImplementation Dependencies.androidx_test_junit
